requirements.txt
fastapi~=0.115.11
pydantic~=2.10.6
aioredis~=2.0.1
sqlalchemy~=2.0.39
uvicorn~=0.34.0
redis~=5.2.1
python-dotenv~=1.0.1
alembic~=1.14.1
pydantic~=2.10.6
psycopg2-binary>=2.9.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0